fix: boost::any_cast<size_t> → boost::any_cast<int> for extruders_count
extruders_count config option has type coInt, so the underlying boost::any
stores an int. Casting it as size_t directly throws boost::bad_any_cast,
which wxWidgets reports as "Unhandled unknown exception; terminating the
application." Use the safe pattern already established at line 4688:
cast to int first, then convert to size_t.

+++ src/slic3r/GUI/Tab.cpp
@@ -1839,7 +1839,7 @@ void Tab::on_value_change(const std::string& opt_key, const boost::any& value)
 
     //Orca: sync filament num if it's a multi tool printer
     if (opt_key == "extruders_count" && !m_config->opt_bool("single_extruder_multi_material")){
-        auto num_extruder = boost::any_cast<size_t>(value);
+        auto num_extruder = size_t(boost::any_cast<int>(value));
         int         old_filament_size = wxGetApp().preset_bundle->filament_presets.size();
         std::vector<std::string> new_colors;
         for (int i = old_filament_size; i < num_extruder; ++i) {
